Add game and author functionalities
