+++ author.json
@@ -1,0 +1,1 @@
+[{"first_name":"John", "last_name": "doe"}, {"first_name":"Mary", "last_name": "Jean"}]
