+++ author.rb
@@ -1,16 +1,16 @@
 class Author
-    attr_accessor :first_name, :last_name,
-    attr_reader :id, :items
+  attr_accessor :first_name, :last_name, :id
+  attr_reader :items
 
-    def initialize(first_name, last_name, items)
-        @id = Random.rand(1..1000)
-        @first_name = first_name
-        @last_name = last_name
-        @items = []
-      end
+  def initialize(first_name, last_name)
+    @id = Random.rand(1..1000)
+    @first_name = first_name
+    @last_name = last_name
+    @items = []
+  end
 
-    def add_item(item)
-        @items.push(item)
-        item.author = self
-    end
+  def add_item(item)
+    @items.push(item)
+    item.author = self
+  end
 end
