+++ book.rb
@@ -10,6 +10,6 @@ class Book < Item
   end
 
   def can_be_archived?
-    super || @cover_state == "bad"
+    super || @cover_state == 'bad'
   end
 end
