+++ game.json
@@ -1,0 +1,1 @@
+[{"multiplayer":"true","last_played_at":"2020-01-1"},{"multiplayer":"True","last_played_at":"2222/2/2"},{"multiplayer":"False","last_played_at":"2222/2/2"}]
