+++ item.rb
@@ -1,5 +1,5 @@
 class Item
-  attr_accessor :name, :description, :type, :published_date, :archived, :tags, :images
+  attr_accessor :name, :description, :type, :published_date, :archived, :tags, :images, :author
 
   def initialize(name, description, type, published_date)
     @name = name
@@ -9,13 +9,18 @@ class Item
     @archived = false
   end
 
+  def add_author(author)
+    @author = author
+    author.items.push(self) unless author.items.include?(self)
+  end
+
   def can_be_archived?
     (Time.now.year - @published_date.year) > 10
   end
 
   def move_to_archive
-    if can_be_archived?
-      @archived = true
-    end
+    return unless can_be_archived?
+
+    @archived = true
   end
 end
